refactoring: fix PR comments
