+++ commons/src/main/java/com/jeanpaulo/musiclibrary/commons/VersionCheck.kt
@@ -1,0 +1,5 @@
+package com.jeanpaulo.musiclibrary.commons
+
+import android.os.Build
+
+fun isAtLeastVersionS() = Build.VERSION.SDK_INT >= Build.VERSION_CODES.S
